libnvmf: Checkpoint further towards TCP transport.
Flesh out hook for receiving data from a received capsule command.
This includes parsing received command capsule PDUs and handling ICD.
This also includes sending an R2T, but does not yet handle parsing
received C2H data PDUs.

+++ lib/libnvmf/internal.h
@@ -51,7 +51,7 @@ struct nvmf_transport_ops {
 
 	/* Receiving controller data. */
 	int (*receive_controller_data)(struct nvmf_capsule *nc,
-	    off_t data_offset, struct iovec *iov, u_int iovcnt);
+	    uint32_t data_offset, struct iovec *iov, u_int iovcnt);
 };
 
 struct nvmf_connection {
